pyproject.toml
[tool.poetry]
name = "moody-backend"
version = "0.1.0"
description = ""
authors = ["SkazuHD <voda0110@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.12"}
transformers = "^4.51.3"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
torch = "^2.7.0"
torchvision = "^0.22.0"
torchaudio = "^2.7.0"
groq = "^0.25.0"
pydantic-settings = "^2.9.1"
python-dotenv = "^1.1.0"
huggingface-hub = "^0.32.2"
soundfile = "^0.13.1"
librosa = "^0.11.0"
pydub = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
